Updates RPKManager configuration dictionary to use modern keys
Aligns code creating the configuration dictionary in the example app
with documentation.
Also updates `README.md` to suggest using the user's PK Kit URL and
Token values.
I'm a little bummed that, out of the box, the app simply crashes. But
absent `#warning` in Objective C, my only other recourse would be to
catch the `Invalid ProximityKit Configuration` Objective C exception in Swift.  But that'd pollute the code
base with some ugly Objective C code that would no longer be needed
once the user sets up the configuration properly.
So... here we are.